Make /register an alias for /user rather than a redirect
